Search by hashtag added
Users can now click on a hashtag link on any given thought post to see a view containing all thoughts containing the selected hashtag. This is the core functionality of the hashtag feature for this web app.

Additionally: fixed a bug where maxlength restrictions weren't being enforced on 'name' and 'thought' attributes of the ThoughtSchema.